requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
requests==2.31.0
Pillow
psycopg2-binary>=2.9
yookassa>=3.0.0
aiohttp>=3.9.0

# Админка (Starlette на aiohttp по /admin)
starlette>=0.36.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.0.0
aiohttp-asgi>=0.5.0